// src/app/context/api/time-attendance/worksheet.model.ts
export type IWorksheet = {
    empID: number;
    empCd: string;
    empName: string;
    deptId: number;
    deptCd: string;
    deptDs: string;
    trnDate: Date;
    timeIn: Date;
    timeOut: Date;
    nh: number;
    oT1: number;
    oT2: number;
    oT3: number;
    missTime: number;
    brkHrs: number;
    lateEntryHrs: number;
    earlyExitHrs: number;
    reason: number;
    reasonName: string;
    multipleReason: string;
    rColor: string;
    edited: boolean;
    editedBy: number;
    editedOn: Date;
    remarks: string;
    empRemarks: null;
    wr: number;
    shift: number;
    wrName: string;
    shiftName: string;
    shiftIN: Date;
    shiftOUT: Date;
    shiftOutNextDay: boolean;
    offDay: boolean;
    holiday: boolean;
    nextDay: boolean;
    approved: boolean;
    leExcusedHrs: number;
    eeExcusedHrs: number;
    eeExcuseReason: number;
    leExcuseReason: number;
    eeExcuseReasonDs: string;
    leExcuseReasonDs: string;
    graceLE: number;
    graceEE: number;
    isSpecialLeaveInHrs: boolean;
    specialLeaveHrType: null;
}